requirements.txt
setuptools~=65.5.1
rdflib~=6.3.2
neo4j~=5.10.0
testcontainers~=3.7.1
pytest~=7.4.0
python-dotenv~=1.0.0